Add test for app:scaffold_data command
